requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
aiohttp==3.11.16
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.8.1
asgi-correlation-id==4.3.4

azure-cosmos==4.9.0
azure-identity==1.21.0
azure-monitor-opentelemetry==1.6.5

pytest==8.3.5
pytest-asyncio==0.26.0
